LayoutTests/imported/w3c/web-platform-tests/service-workers/stub-4.6.3-cache-storage-expected.txt

CONSOLE MESSAGE: line 871: Invalid object member name CacheStorageIterationCallback

Harness Error (FAIL), message = Invalid object member name CacheStorageIterationCallback

PASS CacheStorage interface: existence and properties of interface object 
FAIL CacheStorage interface object length assert_equals: wrong value for CacheStorage.length expected 1 but got 0
PASS CacheStorage interface object name 
PASS CacheStorage interface: existence and properties of interface prototype object 
PASS CacheStorage interface: existence and properties of interface prototype object's "constructor" property 
PASS CacheStorage interface: existence and properties of interface prototype object's @@unscopables property 
PASS CacheStorage interface: operation match(ScalarValueString, DOMString) 
FAIL CacheStorage interface: operation get(DOMString) assert_own_property: interface prototype object missing non-static operation expected property "get" missing
PASS CacheStorage interface: operation has(DOMString) 
FAIL CacheStorage interface: operation set(DOMString, Cache) assert_own_property: interface prototype object missing non-static operation expected property "set" missing
FAIL CacheStorage interface: operation clear() assert_own_property: interface prototype object missing non-static operation expected property "clear" missing
PASS CacheStorage interface: operation delete(DOMString) 
FAIL CacheStorage interface: operation forEach(CacheStorageIterationCallback, object) assert_own_property: interface prototype object missing non-static operation expected property "forEach" missing
FAIL CacheStorage interface: operation entries() assert_own_property: interface prototype object missing non-static operation expected property "entries" missing
PASS CacheStorage interface: operation keys() 
FAIL CacheStorage interface: operation values() assert_own_property: interface prototype object missing non-static operation expected property "values" missing
FAIL CacheStorage interface: operation size() assert_own_property: interface prototype object missing non-static operation expected property "size" missing
FAIL CacheStorage must be primary interface of throw new Error ('No object defined for the CacheStorage interface') assert_equals: Unexpected exception when evaluating object expected null but got object "Error: No object defined for the CacheStorage interface"
FAIL Stringification of throw new Error ('No object defined for the CacheStorage interface') assert_equals: Unexpected exception when evaluating object expected null but got object "Error: No object defined for the CacheStorage interface"
FAIL CacheStorage interface: throw new Error ('No object defined for the CacheStorage interface') must inherit property "match(ScalarValueString, DOMString)" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "Error: No object defined for the CacheStorage interface"
FAIL CacheStorage interface: calling match(ScalarValueString, DOMString) on throw new Error ('No object defined for the CacheStorage interface') with too few arguments must throw TypeError assert_equals: Unexpected exception when evaluating object expected null but got object "Error: No object defined for the CacheStorage interface"
FAIL CacheStorage interface: throw new Error ('No object defined for the CacheStorage interface') must inherit property "get(DOMString)" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "Error: No object defined for the CacheStorage interface"
FAIL CacheStorage interface: calling get(DOMString) on throw new Error ('No object defined for the CacheStorage interface') with too few arguments must throw TypeError assert_equals: Unexpected exception when evaluating object expected null but got object "Error: No object defined for the CacheStorage interface"
FAIL CacheStorage interface: throw new Error ('No object defined for the CacheStorage interface') must inherit property "has(DOMString)" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "Error: No object defined for the CacheStorage interface"
FAIL CacheStorage interface: calling has(DOMString) on throw new Error ('No object defined for the CacheStorage interface') with too few arguments must throw TypeError assert_equals: Unexpected exception when evaluating object expected null but got object "Error: No object defined for the CacheStorage interface"
FAIL CacheStorage interface: throw new Error ('No object defined for the CacheStorage interface') must inherit property "set(DOMString, Cache)" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "Error: No object defined for the CacheStorage interface"
FAIL CacheStorage interface: calling set(DOMString, Cache) on throw new Error ('No object defined for the CacheStorage interface') with too few arguments must throw TypeError assert_equals: Unexpected exception when evaluating object expected null but got object "Error: No object defined for the CacheStorage interface"
FAIL CacheStorage interface: throw new Error ('No object defined for the CacheStorage interface') must inherit property "clear()" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "Error: No object defined for the CacheStorage interface"
FAIL CacheStorage interface: throw new Error ('No object defined for the CacheStorage interface') must inherit property "delete(DOMString)" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "Error: No object defined for the CacheStorage interface"
FAIL CacheStorage interface: calling delete(DOMString) on throw new Error ('No object defined for the CacheStorage interface') with too few arguments must throw TypeError assert_equals: Unexpected exception when evaluating object expected null but got object "Error: No object defined for the CacheStorage interface"
FAIL CacheStorage interface: throw new Error ('No object defined for the CacheStorage interface') must inherit property "forEach(CacheStorageIterationCallback, object)" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "Error: No object defined for the CacheStorage interface"
FAIL CacheStorage interface: calling forEach(CacheStorageIterationCallback, object) on throw new Error ('No object defined for the CacheStorage interface') with too few arguments must throw TypeError assert_equals: Unexpected exception when evaluating object expected null but got object "Error: No object defined for the CacheStorage interface"
FAIL CacheStorage interface: throw new Error ('No object defined for the CacheStorage interface') must inherit property "entries()" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "Error: No object defined for the CacheStorage interface"
FAIL CacheStorage interface: throw new Error ('No object defined for the CacheStorage interface') must inherit property "keys()" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "Error: No object defined for the CacheStorage interface"
FAIL CacheStorage interface: throw new Error ('No object defined for the CacheStorage interface') must inherit property "values()" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "Error: No object defined for the CacheStorage interface"
FAIL CacheStorage interface: throw new Error ('No object defined for the CacheStorage interface') must inherit property "size()" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "Error: No object defined for the CacheStorage interface"
